Applied current production patch

+++ theme/_js/general.js
@@ -1,7 +1,7 @@
 /* -- DEBUG -- */
 if (!navigator.userAgent.match(/Journal/))
 {
-	if (typeof DEBUG_MODE != 'undefined' && DEBUG_MODE) {
+	if (typeof DEBUG_MODE != 'undefined' && DEBUG_MODE && false) {
 		localStorage.clear(); 
 		var p = "/themes/publication_";
 		destroyLessCache(p + "1/theme_1/_css/");
@@ -57,6 +57,8 @@ $(document).ready(function() {
 	updateScreenMode();	
 	
 	//collapsableElements();
+
+	$('.weitere a').click(loadWeitereArtikel);
 
 	scrollToArticleTop();
